fix(dto)
fix dto format

+++ src/main/java/com/progra3/cafeteria_api/model/dto/CustomerResponseDTO.java
@@ -10,5 +10,5 @@ public record CustomerResponseDTO (
     String dni,
     String phoneNumber,
     String email,
-    Double discount
+    Integer discount
 ){}
